Add: simple live filter for results images, filter target is image tooltip Change: fav/tag indicator move to bottom/right Change: supported return range of similarity, but maybe reduce UI response and increment system resource usage fixed: some UI display minor bugs

+++ ImageSearch/settings.cs
@@ -72,7 +72,7 @@ namespace ImageSearch
 
         public bool AllFolder { get; set; } = true;
 
-        public int ResultLimit { get; set; } = 10;
+        public double ResultLimit { get; set; } = 10;
         public double ResultConfidence { get; set; } = 0.25;
 
         public string LastImageFolder { get; set; } = string.Empty;
